# contract/openapi/khaikang.v1.yaml
openapi: 3.1.1
info:
  title: KhaiKang API
  version: 1.0.0
  description: Public HTTP contract for KhaiKang v1.
paths:
...
  /api/v1/projects/{projectId}/issues/{issueId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: issueId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Issues]
      operationId: GetIssue
      security:
        - sessionCookie: []
      responses:
        '200':
          description: A single task visible to the current project member.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '401':
          description: No valid authenticated session exists.
        '404':
          description: The project or task does not exist or is not visible.
    put:
      tags: [Issues]
      operationId: UpdateIssue
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIssueRequest'
      responses:
        '200':
          description: The task details were updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '400':
          description: Request validation or option selection failed.
        '401':
          description: No valid authenticated session exists.
        '403':
          description: The current member lacks issue.update.
        '404':
          description: The project or task does not exist or is not visible.
        '409':
          description: The project is inactive, or the task was updated by another user.
...
  /api/v1/projects/{projectId}/issues/{issueId}/attachments:
    parameters:
      - { name: projectId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: issueId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Issues]
      operationId: ListIssueAttachments
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Active attachments linked to the issue.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IssueAttachmentResponse' }
        '404': { description: The project or issue is not visible. }
    post:
      tags: [Issues]
      operationId: UploadIssueAttachment
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: Attachment stored and linked to the issue.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IssueAttachmentResponse' }
        '400': { description: The file is empty or invalid. }
        '403': { description: The current member lacks issue.attachment.upload. }
        '404': { description: The project or issue does not exist. }
        '409': { description: The project is inactive. }
        '413': { description: The file exceeds the configured size limit. }
  /api/v1/projects/{projectId}/issues/{issueId}/attachments/{attachmentId}:
    parameters:
      - { name: projectId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: issueId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: attachmentId, in: path, required: true, schema: { type: string, format: uuid } }
    delete:
      tags: [Issues]
      operationId: DeleteIssueAttachment
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '204': { description: The attachment link was soft deleted. }
        '403': { description: The current member lacks issue.attachment.delete. }
        '404': { description: The attachment does not exist or was already removed. }
        '409': { description: The project is inactive. }
  /api/v1/projects/{projectId}/issues/{issueId}/attachments/{attachmentId}/content:
    parameters:
      - { name: projectId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: issueId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: attachmentId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: inline, in: query, required: false, schema: { type: boolean, default: false } }
    get:
      tags: [Issues]
      operationId: GetIssueAttachmentContent
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Attachment bytes. Images may be rendered inline when requested.
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '403': { description: The current member lacks issue.read. }
        '404': { description: The attachment does not exist or was removed. }
        '503': { description: The configured storage provider cannot read the file. }
...
  /api/v1/test-workspaces/{workspaceId}/cases/{caseId}:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: caseId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Test Management]
      operationId: GetTestCase
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Test case details visible to the current workspace member.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestCaseResponse' }
        '404': { description: Test case or workspace not found. }
    put:
      tags: [Test Management]
      operationId: UpdateTestCase
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTestCaseRequest' }
      responses:
        '200':
          description: Test case updated with new properties and steps.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestCaseResponse' }
        '400': { description: Request validation failed. }
        '403': { description: Current role cannot manage test cases. }
        '404': { description: Test case, workspace, or suite does not exist. }
        '409': { description: Version conflict or invalid suite target. }
  /api/v1/test-workspaces/{workspaceId}/cases/{caseId}/attachments:
    parameters:
      - { name: workspaceId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: caseId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Test Management]
      operationId: ListTestCaseAttachments
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Active attachments linked to the test case.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/TestCaseAttachmentResponse' } }
        '404': { description: Test case or workspace is not visible. }
    post:
      tags: [Test Management]
      operationId: UploadTestCaseAttachment
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: Attachment stored and linked to the test case.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestCaseAttachmentResponse' }
        '403': { description: Current role cannot manage test case attachments. }
        '404': { description: Test case or workspace not found. }
        '413': { description: File exceeds the configured size limit. }
  /api/v1/test-workspaces/{workspaceId}/cases/{caseId}/attachments/{attachmentId}:
    parameters:
      - { name: workspaceId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: caseId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: attachmentId, in: path, required: true, schema: { type: string, format: uuid } }
    delete:
      tags: [Test Management]
      operationId: DeleteTestCaseAttachment
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '204': { description: Attachment soft-deleted. }
        '403': { description: Current role cannot manage test case attachments. }
        '404': { description: Attachment, test case, or workspace not found. }
  /api/v1/test-workspaces/{workspaceId}/cases/{caseId}/attachments/{attachmentId}/content:
    parameters:
      - { name: workspaceId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: caseId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: attachmentId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: inline, in: query, required: false, schema: { type: boolean, default: false } }
    get:
      tags: [Test Management]
      operationId: GetTestCaseAttachmentContent
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Attachment bytes; images may be rendered inline.
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '404': { description: Attachment, test case, or workspace not found. }
        '503': { description: Attachment storage is unavailable. }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: KhaiKang.Session
  parameters:
...
    CsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: true
      description: Token returned by `getCsrfToken`; its paired cookie is also required.
      schema:
        type: string
  schemas:
...
    CreateTestCaseStepRequest:
      type: object
      additionalProperties: false
      required: [action, expectedResult]
      properties:
        action: { type: string, minLength: 1, maxLength: 4000 }
        expectedResult: { type: string, minLength: 1, maxLength: 4000 }
...
    UpdateTestCaseRequest:
      type: object
      additionalProperties: false
      required:
        - suiteId
        - title
        - description
        - preconditions
        - overallExpectedResult
        - sortOrder
        - status
        - version
        - steps
      properties:
        suiteId: { type: string, format: uuid }
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: [string, 'null'], maxLength: 4000 }
        preconditions: { type: [string, 'null'], maxLength: 4000 }
        overallExpectedResult: { type: [string, 'null'], maxLength: 4000 }
        sortOrder: { type: integer, minimum: 0 }
        status: { type: string, enum: [active, inactive] }
        version: { type: integer, minimum: 1 }
        tagIds:
          type: array
          uniqueItems: true
          items: { type: string, format: uuid }
        steps:
          type: array
          minItems: 1
          maxItems: 100
          items: { $ref: '#/components/schemas/CreateTestCaseStepRequest' }
    TestCaseStepResponse:
      type: object
      additionalProperties: false
      required: [id, stepNo, action, expectedResult]
      properties:
        id: { type: string, format: uuid }
        stepNo: { type: integer, minimum: 1 }
        action: { type: string }
        expectedResult: { type: string }
    TestCaseResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - suiteId
        - caseNo
        - tags
        - title
        - description
        - preconditions
        - overallExpectedResult
        - sortOrder
        - status
        - steps
        - createdAt
        - updatedAt
        - version
      properties:
        id: { type: string, format: uuid }
        suiteId: { type: string, format: uuid }
        caseNo: { type: integer, minimum: 1 }
        tags:
          type: array
          items: { $ref: '#/components/schemas/TestTagResponse' }
        title: { type: string }
        description: { type: [string, 'null'] }
        preconditions: { type: [string, 'null'] }
        overallExpectedResult: { type: [string, 'null'] }
        sortOrder: { type: integer }
        status: { type: string, enum: [active, inactive] }
        steps:
          type: array
          items: { $ref: '#/components/schemas/TestCaseStepResponse' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: integer, minimum: 1 }
    TestCaseAttachmentResponse:
      type: object
      additionalProperties: false
      required: [id, testCaseId, originalFileName, contentType, fileSize, fileHash, uploadedByAccountId, uploadedByUsername, createdAt]
      properties:
        id: { type: string, format: uuid }
        testCaseId: { type: string, format: uuid }
        originalFileName: { type: string, maxLength: 255 }
        contentType: { type: string, maxLength: 200 }
        fileSize: { type: integer, format: int64, minimum: 1 }
        fileHash: { type: string, minLength: 64, maxLength: 64 }
        uploadedByAccountId: { type: string, format: uuid }
        uploadedByUsername: { type: string }
        createdAt: { type: string, format: date-time }
    TestTagResponse:
      type: object
      additionalProperties: false
      required: [id, name, description, status, version]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, maxLength: 50 }
        description: { type: [string, 'null'], maxLength: 4000 }
        status: { type: string, enum: [active, inactive] }
        version: { type: integer, minimum: 1 }
...
    UpdateIssueRequest:
      type: object
      additionalProperties: false
      required: [title, typeCode, priorityCode, version]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        typeCode:
          type: string
        priorityCode:
          type: string
        description:
          type: [string, 'null']
          maxLength: 20000
        userStory:
          type: [string, 'null']
          maxLength: 20000
        definitionOfDone:
          type: [string, 'null']
          maxLength: 20000
        completionSummary:
          type: [string, 'null']
          maxLength: 20000
        version:
          type: integer
          format: int32
          minimum: 1
...
    IssueResponse:
      type: object
      additionalProperties: false
      required: [id, projectId, issueNo, key, title, typeCode, typeName, statusCode, statusName, priorityCode, priorityName, reporterAccountId, reporterUsername, createdAt, updatedAt, version]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        issueNo:
          type: integer
          format: int32
        key:
          type: string
        title:
          type: string
        description:
          type: [string, 'null']
        userStory:
          type: [string, 'null']
        definitionOfDone:
          type: [string, 'null']
        completionSummary:
          type: [string, 'null']
        typeCode:
          type: string
        typeName:
          type: string
        statusCode:
          type: string
        statusName:
          type: string
        priorityCode:
          type: string
        priorityName:
          type: string
        reporterAccountId:
          type: string
          format: uuid
        reporterUsername:
          type: string
        assigneeAccountId:
          type: [string, 'null']
          format: uuid
        assigneeUsername:
          type: [string, 'null']
        completedAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int32
    IssueAttachmentResponse:
      type: object
      additionalProperties: false
      required: [id, issueId, originalFileName, contentType, fileSize, fileHash, uploadedByAccountId, uploadedByUsername, createdAt]
      properties:
        id: { type: string, format: uuid }
        issueId: { type: string, format: uuid }
        originalFileName: { type: string, maxLength: 255 }
        contentType: { type: string, maxLength: 200 }
        fileSize: { type: integer, format: int64, minimum: 1 }
        fileHash: { type: string, minLength: 64, maxLength: 64 }
        uploadedByAccountId: { type: string, format: uuid }
        uploadedByUsername: { type: string }
        createdAt: { type: string, format: date-time }
...